- remove compilation warnings
git-svn-id: https://svn.code.sf.net/p/auber/code/tulip@4780 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip/src/GraphMeasure.cpp
@@ -93,7 +93,7 @@ double tlp::averagePathLength(const Graph *graph,
 #ifdef _OPENMP
 #pragma omp parallel for private(i) schedule(dynamic, 1)
 #endif
-  for (i = 0; i < nbNodes; ++i) {
+  for (i = 0; i < static_cast<int>(nbNodes); ++i) {
     if (stopfor) continue;
     // check if we are in the main thread.
     // Qt calls can't be done outside of the master thread
